chore: rename client to repository
